move the logic to execute classdef statements into ClassDef.run
